@@ -7,7 +7,7 @@ name = "fltk"
 dynamic = []
 description = 'Formal Language ToolKit'
 version = "0.0.1"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 license = "MIT"
 keywords = []
 authors = [
@@ -16,10 +16,9 @@ authors = [
 classifiers = [
   "Development Status :: 4 - Beta",
   "Programming Language :: Python",
-  "Programming Language :: Python :: 3.8",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: Implementation :: CPython",
   "Programming Language :: Python :: Implementation :: PyPy",
 ]
@@ -44,18 +43,17 @@ lint = [
 ]
 
 [tool.pyright]
-pythonVersion = "3.8"
+pythonVersion = "3.10"
 include = ["fltk", "tests"]
 exclude = ["**/__pycache__"]
 stubPath = ""
 venvPath = ".venv"
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py310"
 line-length = 120
 
 [tool.ruff.format]
-quote-style = "single"
 skip-magic-trailing-comma = false
 
 [tool.ruff.lint]
