@@ -62,7 +62,7 @@ class Parser:
 
     def consume_literal(
         self, pos: int, literal: str
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         if span := self.terminalsrc.consume_literal(pos=pos, literal=literal):
             return fltk.fegen.pyrt.memo.ApplyResult(pos=span.end, result=span)
         self.error_tracker.fail_literal(pos=pos, rule_id=self.packrat.invocation_stack[-1], literal=literal)
@@ -70,27 +70,23 @@ class Parser:
 
     def consume_regex(
         self, pos: int, regex: str
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         if span := self.terminalsrc.consume_regex(pos=pos, regex=regex):
             return fltk.fegen.pyrt.memo.ApplyResult(pos=span.end, result=span)
         self.error_tracker.fail_regex(pos=pos, rule_id=self.packrat.invocation_stack[-1], regex=regex)
         return None
 
-    def parse_grammar(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Grammar]]:
+    def parse_grammar(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Grammar] | None:
         if alt0 := self.parse_grammar__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_grammar(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Grammar]]:
+    def apply__parse_grammar(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Grammar] | None:
         return self.packrat.apply(
             rule_callable=self.parse_grammar, rule_id=0, rule_cache=self._cache__parse_grammar, pos=pos
         )
 
-    def parse_grammar__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Grammar]]:
+    def parse_grammar__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Grammar] | None:
         result: bootstrap_cst.Grammar = bootstrap_cst.Grammar(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_grammar__alt0__item0(pos=pos):
             pos = item0.pos
@@ -102,7 +98,7 @@ class Parser:
 
     def parse_grammar__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Grammar]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Grammar] | None:
         result: bootstrap_cst.Grammar = bootstrap_cst.Grammar(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         while one_result := self.apply__parse_rule(pos=pos):
             pos = one_result.pos
@@ -112,15 +108,15 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_rule(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Rule]]:
+    def parse_rule(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Rule] | None:
         if alt0 := self.parse_rule__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_rule(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Rule]]:
+    def apply__parse_rule(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Rule] | None:
         return self.packrat.apply(rule_callable=self.parse_rule, rule_id=1, rule_cache=self._cache__parse_rule, pos=pos)
 
-    def parse_rule__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Rule]]:
+    def parse_rule__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Rule] | None:
         result: bootstrap_cst.Rule = bootstrap_cst.Rule(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_rule__alt0__item0(pos=pos):
             pos = item0.pos
@@ -153,41 +149,39 @@ class Parser:
 
     def parse_rule__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier] | None:
         return self.apply__parse_identifier(pos=pos)
 
     def parse_rule__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=":=")
 
     def parse_rule__alt0__item2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives] | None:
         return self.apply__parse_alternatives(pos=pos)
 
     def parse_rule__alt0__item3(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=";")
 
-    def parse_alternatives(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives]]:
+    def parse_alternatives(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives] | None:
         if alt0 := self.parse_alternatives__alt0(pos=pos):
             return alt0
         return None
 
     def apply__parse_alternatives(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives] | None:
         return self.packrat.apply(
             rule_callable=self.parse_alternatives, rule_id=2, rule_cache=self._cache__parse_alternatives, pos=pos
         )
 
     def parse_alternatives__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives] | None:
         result: bootstrap_cst.Alternatives = bootstrap_cst.Alternatives(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -206,19 +200,19 @@ class Parser:
 
     def parse_alternatives__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         return self.apply__parse_items(pos=pos)
 
     def parse_alternatives__alt0__item1__alts(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives] | None:
         if alt0 := self.parse_alternatives__alt0__item1__alts__alt0(pos=pos):
             return alt0
         return None
 
     def parse_alternatives__alt0__item1__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives] | None:
         result: bootstrap_cst.Alternatives = bootstrap_cst.Alternatives(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -238,17 +232,17 @@ class Parser:
 
     def parse_alternatives__alt0__item1__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="|")
 
     def parse_alternatives__alt0__item1__alts__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         return self.apply__parse_items(pos=pos)
 
     def parse_alternatives__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives] | None:
         result: bootstrap_cst.Alternatives = bootstrap_cst.Alternatives(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -258,21 +252,17 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_items(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    def parse_items(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         if alt0 := self.parse_items__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_items(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    def apply__parse_items(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         return self.packrat.apply(
             rule_callable=self.parse_items, rule_id=3, rule_cache=self._cache__parse_items, pos=pos
         )
 
-    def parse_items__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    def parse_items__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         result: bootstrap_cst.Items = bootstrap_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item0(pos=pos):
             pos = item0.pos
@@ -294,21 +284,19 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_items__alt0__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item]]:
+    def parse_items__alt0__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item] | None:
         return self.apply__parse_item(pos=pos)
 
     def parse_items__alt0__item1__alts(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         if alt0 := self.parse_items__alt0__item1__alts__alt0(pos=pos):
             return alt0
         return None
 
     def parse_items__alt0__item1__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         result: bootstrap_cst.Items = bootstrap_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item1__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -327,7 +315,7 @@ class Parser:
 
     def parse_items__alt0__item1__alts__alt0__item0__alts(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         if alt0 := self.parse_items__alt0__item1__alts__alt0__item0__alts__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_items__alt0__item1__alts__alt0__item0__alts__alt1(pos=pos):
@@ -336,7 +324,7 @@ class Parser:
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         result: bootstrap_cst.Items = bootstrap_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item1__alts__alt0__item0__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -348,12 +336,12 @@ class Parser:
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=".")
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         result: bootstrap_cst.Items = bootstrap_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item1__alts__alt0__item0__alts__alt1__item0(pos=pos):
             pos = item0.pos
@@ -365,22 +353,20 @@ class Parser:
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt1__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=",")
 
     def parse_items__alt0__item1__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         return self.parse_items__alt0__item1__alts__alt0__item0__alts(pos=pos)
 
     def parse_items__alt0__item1__alts__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item] | None:
         return self.apply__parse_item(pos=pos)
 
-    def parse_items__alt0__item1(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    def parse_items__alt0__item1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         result: bootstrap_cst.Items = bootstrap_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         while one_result := self.parse_items__alt0__item1__alts(pos=pos):
             pos = one_result.pos
@@ -390,7 +376,7 @@ class Parser:
 
     def parse_items__alt0__item2__alts(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         if alt0 := self.parse_items__alt0__item2__alts__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_items__alt0__item2__alts__alt1(pos=pos):
@@ -399,7 +385,7 @@ class Parser:
 
     def parse_items__alt0__item2__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         result: bootstrap_cst.Items = bootstrap_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item2__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -411,12 +397,12 @@ class Parser:
 
     def parse_items__alt0__item2__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=".")
 
     def parse_items__alt0__item2__alts__alt1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         result: bootstrap_cst.Items = bootstrap_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item2__alts__alt1__item0(pos=pos):
             pos = item0.pos
@@ -428,23 +414,21 @@ class Parser:
 
     def parse_items__alt0__item2__alts__alt1__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=",")
 
-    def parse_items__alt0__item2(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items]]:
+    def parse_items__alt0__item2(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Items] | None:
         return self.parse_items__alt0__item2__alts(pos=pos)
 
-    def parse_item(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item]]:
+    def parse_item(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item] | None:
         if alt0 := self.parse_item__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_item(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item]]:
+    def apply__parse_item(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item] | None:
         return self.packrat.apply(rule_callable=self.parse_item, rule_id=4, rule_cache=self._cache__parse_item, pos=pos)
 
-    def parse_item__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item]]:
+    def parse_item__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item] | None:
         result: bootstrap_cst.Item = bootstrap_cst.Item(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_item__alt0__item0(pos=pos):
             pos = item0.pos
@@ -467,14 +451,14 @@ class Parser:
 
     def parse_item__alt0__item0__alts(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item] | None:
         if alt0 := self.parse_item__alt0__item0__alts__alt0(pos=pos):
             return alt0
         return None
 
     def parse_item__alt0__item0__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item] | None:
         result: bootstrap_cst.Item = bootstrap_cst.Item(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_item__alt0__item0__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -490,35 +474,31 @@ class Parser:
 
     def parse_item__alt0__item0__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier] | None:
         return self.apply__parse_identifier(pos=pos)
 
     def parse_item__alt0__item0__alts__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=":")
 
-    def parse_item__alt0__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item]]:
+    def parse_item__alt0__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Item] | None:
         return self.parse_item__alt0__item0__alts(pos=pos)
 
     def parse_item__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition] | None:
         return self.apply__parse_disposition(pos=pos)
 
-    def parse_item__alt0__item2(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term]]:
+    def parse_item__alt0__item2(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term] | None:
         return self.apply__parse_term(pos=pos)
 
     def parse_item__alt0__item3(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier] | None:
         return self.apply__parse_quantifier(pos=pos)
 
-    def parse_term(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term]]:
+    def parse_term(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term] | None:
         if alt0 := self.parse_term__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_term__alt1(pos=pos):
@@ -529,10 +509,10 @@ class Parser:
             return alt3
         return None
 
-    def apply__parse_term(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term]]:
+    def apply__parse_term(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term] | None:
         return self.packrat.apply(rule_callable=self.parse_term, rule_id=5, rule_cache=self._cache__parse_term, pos=pos)
 
-    def parse_term__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term]]:
+    def parse_term__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term] | None:
         result: bootstrap_cst.Term = bootstrap_cst.Term(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_term__alt0__item0(pos=pos):
             pos = item0.pos
@@ -544,10 +524,10 @@ class Parser:
 
     def parse_term__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier] | None:
         return self.apply__parse_identifier(pos=pos)
 
-    def parse_term__alt1(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term]]:
+    def parse_term__alt1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term] | None:
         result: bootstrap_cst.Term = bootstrap_cst.Term(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_term__alt1__item0(pos=pos):
             pos = item0.pos
@@ -557,12 +537,10 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_term__alt1__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Literal]]:
+    def parse_term__alt1__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Literal] | None:
         return self.apply__parse_literal(pos=pos)
 
-    def parse_term__alt2(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term]]:
+    def parse_term__alt2(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term] | None:
         result: bootstrap_cst.Term = bootstrap_cst.Term(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_term__alt2__item0(pos=pos):
             pos = item0.pos
@@ -582,20 +560,20 @@ class Parser:
 
     def parse_term__alt2__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="/")
 
     def parse_term__alt2__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.RawString]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.RawString] | None:
         return self.apply__parse_raw_string(pos=pos)
 
     def parse_term__alt2__item2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="/")
 
-    def parse_term__alt3(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term]]:
+    def parse_term__alt3(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Term] | None:
         result: bootstrap_cst.Term = bootstrap_cst.Term(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_term__alt3__item0(pos=pos):
             pos = item0.pos
@@ -619,22 +597,20 @@ class Parser:
 
     def parse_term__alt3__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="(")
 
     def parse_term__alt3__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Alternatives] | None:
         return self.apply__parse_alternatives(pos=pos)
 
     def parse_term__alt3__item2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=")")
 
-    def parse_disposition(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition]]:
+    def parse_disposition(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition] | None:
         if alt0 := self.parse_disposition__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_disposition__alt1(pos=pos):
@@ -645,14 +621,14 @@ class Parser:
 
     def apply__parse_disposition(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition] | None:
         return self.packrat.apply(
             rule_callable=self.parse_disposition, rule_id=6, rule_cache=self._cache__parse_disposition, pos=pos
         )
 
     def parse_disposition__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition] | None:
         result: bootstrap_cst.Disposition = bootstrap_cst.Disposition(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -666,12 +642,12 @@ class Parser:
 
     def parse_disposition__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="%")
 
     def parse_disposition__alt1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition] | None:
         result: bootstrap_cst.Disposition = bootstrap_cst.Disposition(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -685,12 +661,12 @@ class Parser:
 
     def parse_disposition__alt1__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="$")
 
     def parse_disposition__alt2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Disposition] | None:
         result: bootstrap_cst.Disposition = bootstrap_cst.Disposition(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -704,12 +680,10 @@ class Parser:
 
     def parse_disposition__alt2__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="!")
 
-    def parse_quantifier(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier]]:
+    def parse_quantifier(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier] | None:
         if alt0 := self.parse_quantifier__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_quantifier__alt1(pos=pos):
@@ -720,14 +694,14 @@ class Parser:
 
     def apply__parse_quantifier(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier] | None:
         return self.packrat.apply(
             rule_callable=self.parse_quantifier, rule_id=7, rule_cache=self._cache__parse_quantifier, pos=pos
         )
 
     def parse_quantifier__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier] | None:
         result: bootstrap_cst.Quantifier = bootstrap_cst.Quantifier(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -741,12 +715,12 @@ class Parser:
 
     def parse_quantifier__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="?")
 
     def parse_quantifier__alt1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier] | None:
         result: bootstrap_cst.Quantifier = bootstrap_cst.Quantifier(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -760,12 +734,12 @@ class Parser:
 
     def parse_quantifier__alt1__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="+")
 
     def parse_quantifier__alt2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Quantifier] | None:
         result: bootstrap_cst.Quantifier = bootstrap_cst.Quantifier(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -779,26 +753,24 @@ class Parser:
 
     def parse_quantifier__alt2__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="*")
 
-    def parse_identifier(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier]]:
+    def parse_identifier(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier] | None:
         if alt0 := self.parse_identifier__alt0(pos=pos):
             return alt0
         return None
 
     def apply__parse_identifier(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier] | None:
         return self.packrat.apply(
             rule_callable=self.parse_identifier, rule_id=8, rule_cache=self._cache__parse_identifier, pos=pos
         )
 
     def parse_identifier__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Identifier] | None:
         result: bootstrap_cst.Identifier = bootstrap_cst.Identifier(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -812,26 +784,22 @@ class Parser:
 
     def parse_identifier__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_regex(pos=pos, regex="[_a-z][_a-z0-9]*")
 
-    def parse_raw_string(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.RawString]]:
+    def parse_raw_string(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.RawString] | None:
         if alt0 := self.parse_raw_string__alt0(pos=pos):
             return alt0
         return None
 
     def apply__parse_raw_string(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.RawString]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.RawString] | None:
         return self.packrat.apply(
             rule_callable=self.parse_raw_string, rule_id=9, rule_cache=self._cache__parse_raw_string, pos=pos
         )
 
-    def parse_raw_string__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.RawString]]:
+    def parse_raw_string__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.RawString] | None:
         result: bootstrap_cst.RawString = bootstrap_cst.RawString(
             span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1)
         )
@@ -845,24 +813,20 @@ class Parser:
 
     def parse_raw_string__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_regex(pos=pos, regex="([^\\/\\n\\\\]|\\\\.)+")
 
-    def parse_literal(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Literal]]:
+    def parse_literal(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Literal] | None:
         if alt0 := self.parse_literal__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_literal(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Literal]]:
+    def apply__parse_literal(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Literal] | None:
         return self.packrat.apply(
             rule_callable=self.parse_literal, rule_id=10, rule_cache=self._cache__parse_literal, pos=pos
         )
 
-    def parse_literal__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Literal]]:
+    def parse_literal__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, bootstrap_cst.Literal] | None:
         result: bootstrap_cst.Literal = bootstrap_cst.Literal(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_literal__alt0__item0(pos=pos):
             pos = item0.pos
@@ -874,5 +838,5 @@ class Parser:
 
     def parse_literal__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
-        return self.consume_regex(pos=pos, regex='("([^"\\n\\\\]|\\\\.)+"|\'([^\'\\n\\\\]|\\\\.)+\')')
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
+        return self.consume_regex(pos=pos, regex="(\"([^\"\\n\\\\]|\\\\.)+\"|'([^'\\n\\\\]|\\\\.)+')")
