@@ -1,11 +1,10 @@
 import dataclasses as dc
-from typing import Any, Type, TypeVar
+from typing import Any, TypeVar
 
 _T = TypeVar("_T")
-_U = TypeVar("_U")
 
 
-def thaw(cls: Type[_T]) -> Type[_U]:
+def thaw(cls: type[_T]) -> type[Any]:
     fields = dc.fields(cls)  # type: ignore
     result = dc.make_dataclass(
         cls_name=cls.__name__,
@@ -14,5 +13,5 @@ def thaw(cls: Type[_T]) -> Type[_U]:
     return result
 
 
-def as_class(obj: Any, cls: Type[_T]) -> _T:
+def as_class(obj: Any, cls: type[_T]) -> _T:
     return cls(**{f.name: getattr(obj, f.name) for f in dc.fields(obj)})
