@@ -78,7 +78,7 @@ class Parser:
 
     def consume_literal(
         self, pos: int, literal: str
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         if span := self.terminalsrc.consume_literal(pos=pos, literal=literal):
             return fltk.fegen.pyrt.memo.ApplyResult(pos=span.end, result=span)
         self.error_tracker.fail_literal(pos=pos, rule_id=self.packrat.invocation_stack[-1], literal=literal)
@@ -86,25 +86,23 @@ class Parser:
 
     def consume_regex(
         self, pos: int, regex: str
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         if span := self.terminalsrc.consume_regex(pos=pos, regex=regex):
             return fltk.fegen.pyrt.memo.ApplyResult(pos=span.end, result=span)
         self.error_tracker.fail_regex(pos=pos, rule_id=self.packrat.invocation_stack[-1], regex=regex)
         return None
 
-    def parse_grammar(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Grammar]]:
+    def parse_grammar(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Grammar] | None:
         if alt0 := self.parse_grammar__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_grammar(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Grammar]]:
+    def apply__parse_grammar(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Grammar] | None:
         return self.packrat.apply(
             rule_callable=self.parse_grammar, rule_id=0, rule_cache=self._cache__parse_grammar, pos=pos
         )
 
-    def parse_grammar__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Grammar]]:
+    def parse_grammar__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Grammar] | None:
         result: fltk_cst.Grammar = fltk_cst.Grammar(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_grammar__alt0__item0(pos=pos):
             pos = item0.pos
@@ -119,14 +117,10 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_grammar__alt0__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Rule]]:
+    def parse_grammar__alt0__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Rule] | None:
         return self.apply__parse_rule(pos=pos)
 
-    def parse_grammar__alt0__item1(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Grammar]]:
+    def parse_grammar__alt0__item1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Grammar] | None:
         result: fltk_cst.Grammar = fltk_cst.Grammar(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         while one_result := self.apply__parse_rule(pos=pos):
             pos = one_result.pos
@@ -136,15 +130,15 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_rule(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Rule]]:
+    def parse_rule(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Rule] | None:
         if alt0 := self.parse_rule__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_rule(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Rule]]:
+    def apply__parse_rule(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Rule] | None:
         return self.packrat.apply(rule_callable=self.parse_rule, rule_id=1, rule_cache=self._cache__parse_rule, pos=pos)
 
-    def parse_rule__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Rule]]:
+    def parse_rule__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Rule] | None:
         result: fltk_cst.Rule = fltk_cst.Rule(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_rule__alt0__item0(pos=pos):
             pos = item0.pos
@@ -175,43 +169,35 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_rule__alt0__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier]]:
+    def parse_rule__alt0__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier] | None:
         return self.apply__parse_identifier(pos=pos)
 
     def parse_rule__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=":=")
 
-    def parse_rule__alt0__item2(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives]]:
+    def parse_rule__alt0__item2(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives] | None:
         return self.apply__parse_alternatives(pos=pos)
 
     def parse_rule__alt0__item3(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=";")
 
-    def parse_alternatives(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives]]:
+    def parse_alternatives(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives] | None:
         if alt0 := self.parse_alternatives__alt0(pos=pos):
             return alt0
         return None
 
     def apply__parse_alternatives(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives] | None:
         return self.packrat.apply(
             rule_callable=self.parse_alternatives, rule_id=2, rule_cache=self._cache__parse_alternatives, pos=pos
         )
 
-    def parse_alternatives__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives]]:
+    def parse_alternatives__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives] | None:
         result: fltk_cst.Alternatives = fltk_cst.Alternatives(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_alternatives__alt0__item0(pos=pos):
             pos = item0.pos
@@ -226,21 +212,19 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_alternatives__alt0__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    def parse_alternatives__alt0__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         return self.apply__parse_items(pos=pos)
 
     def parse_alternatives__alt0__item1__alts(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives] | None:
         if alt0 := self.parse_alternatives__alt0__item1__alts__alt0(pos=pos):
             return alt0
         return None
 
     def parse_alternatives__alt0__item1__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives] | None:
         result: fltk_cst.Alternatives = fltk_cst.Alternatives(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_alternatives__alt0__item1__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -258,17 +242,17 @@ class Parser:
 
     def parse_alternatives__alt0__item1__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="|")
 
     def parse_alternatives__alt0__item1__alts__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         return self.apply__parse_items(pos=pos)
 
     def parse_alternatives__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives] | None:
         result: fltk_cst.Alternatives = fltk_cst.Alternatives(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         while one_result := self.parse_alternatives__alt0__item1__alts(pos=pos):
             pos = one_result.pos
@@ -276,17 +260,17 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_items(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    def parse_items(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         if alt0 := self.parse_items__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_items(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    def apply__parse_items(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         return self.packrat.apply(
             rule_callable=self.parse_items, rule_id=3, rule_cache=self._cache__parse_items, pos=pos
         )
 
-    def parse_items__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    def parse_items__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         result: fltk_cst.Items = fltk_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item0(pos=pos):
             pos = item0.pos
@@ -308,21 +292,17 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_items__alt0__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item]]:
+    def parse_items__alt0__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item] | None:
         return self.apply__parse_item(pos=pos)
 
-    def parse_items__alt0__item1__alts(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    def parse_items__alt0__item1__alts(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         if alt0 := self.parse_items__alt0__item1__alts__alt0(pos=pos):
             return alt0
         return None
 
     def parse_items__alt0__item1__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         result: fltk_cst.Items = fltk_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item1__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -341,7 +321,7 @@ class Parser:
 
     def parse_items__alt0__item1__alts__alt0__item0__alts(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         if alt0 := self.parse_items__alt0__item1__alts__alt0__item0__alts__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_items__alt0__item1__alts__alt0__item0__alts__alt1(pos=pos):
@@ -352,7 +332,7 @@ class Parser:
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         result: fltk_cst.Items = fltk_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item1__alts__alt0__item0__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -364,12 +344,12 @@ class Parser:
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=".")
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         result: fltk_cst.Items = fltk_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item1__alts__alt0__item0__alts__alt1__item0(pos=pos):
             pos = item0.pos
@@ -381,12 +361,12 @@ class Parser:
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt1__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=",")
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         result: fltk_cst.Items = fltk_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item1__alts__alt0__item0__alts__alt2__item0(pos=pos):
             pos = item0.pos
@@ -398,22 +378,20 @@ class Parser:
 
     def parse_items__alt0__item1__alts__alt0__item0__alts__alt2__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=":")
 
     def parse_items__alt0__item1__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         return self.parse_items__alt0__item1__alts__alt0__item0__alts(pos=pos)
 
     def parse_items__alt0__item1__alts__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item] | None:
         return self.apply__parse_item(pos=pos)
 
-    def parse_items__alt0__item1(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    def parse_items__alt0__item1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         result: fltk_cst.Items = fltk_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         while one_result := self.parse_items__alt0__item1__alts(pos=pos):
             pos = one_result.pos
@@ -421,9 +399,7 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_items__alt0__item2__alts(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    def parse_items__alt0__item2__alts(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         if alt0 := self.parse_items__alt0__item2__alts__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_items__alt0__item2__alts__alt1(pos=pos):
@@ -434,7 +410,7 @@ class Parser:
 
     def parse_items__alt0__item2__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         result: fltk_cst.Items = fltk_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item2__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -446,12 +422,12 @@ class Parser:
 
     def parse_items__alt0__item2__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=".")
 
     def parse_items__alt0__item2__alts__alt1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         result: fltk_cst.Items = fltk_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item2__alts__alt1__item0(pos=pos):
             pos = item0.pos
@@ -463,12 +439,12 @@ class Parser:
 
     def parse_items__alt0__item2__alts__alt1__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=",")
 
     def parse_items__alt0__item2__alts__alt2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         result: fltk_cst.Items = fltk_cst.Items(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_items__alt0__item2__alts__alt2__item0(pos=pos):
             pos = item0.pos
@@ -480,23 +456,21 @@ class Parser:
 
     def parse_items__alt0__item2__alts__alt2__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=":")
 
-    def parse_items__alt0__item2(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items]]:
+    def parse_items__alt0__item2(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Items] | None:
         return self.parse_items__alt0__item2__alts(pos=pos)
 
-    def parse_item(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item]]:
+    def parse_item(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item] | None:
         if alt0 := self.parse_item__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_item(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item]]:
+    def apply__parse_item(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item] | None:
         return self.packrat.apply(rule_callable=self.parse_item, rule_id=4, rule_cache=self._cache__parse_item, pos=pos)
 
-    def parse_item__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item]]:
+    def parse_item__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item] | None:
         result: fltk_cst.Item = fltk_cst.Item(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_item__alt0__item0(pos=pos):
             pos = item0.pos
@@ -517,16 +491,14 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_item__alt0__item0__alts(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item]]:
+    def parse_item__alt0__item0__alts(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item] | None:
         if alt0 := self.parse_item__alt0__item0__alts__alt0(pos=pos):
             return alt0
         return None
 
     def parse_item__alt0__item0__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item] | None:
         result: fltk_cst.Item = fltk_cst.Item(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_item__alt0__item0__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -542,35 +514,27 @@ class Parser:
 
     def parse_item__alt0__item0__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier] | None:
         return self.apply__parse_identifier(pos=pos)
 
     def parse_item__alt0__item0__alts__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=":")
 
-    def parse_item__alt0__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item]]:
+    def parse_item__alt0__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Item] | None:
         return self.parse_item__alt0__item0__alts(pos=pos)
 
-    def parse_item__alt0__item1(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition]]:
+    def parse_item__alt0__item1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition] | None:
         return self.apply__parse_disposition(pos=pos)
 
-    def parse_item__alt0__item2(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term]]:
+    def parse_item__alt0__item2(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term] | None:
         return self.apply__parse_term(pos=pos)
 
-    def parse_item__alt0__item3(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier]]:
+    def parse_item__alt0__item3(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier] | None:
         return self.apply__parse_quantifier(pos=pos)
 
-    def parse_term(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term]]:
+    def parse_term(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term] | None:
         if alt0 := self.parse_term__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_term__alt1(pos=pos):
@@ -581,10 +545,10 @@ class Parser:
             return alt3
         return None
 
-    def apply__parse_term(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term]]:
+    def apply__parse_term(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term] | None:
         return self.packrat.apply(rule_callable=self.parse_term, rule_id=5, rule_cache=self._cache__parse_term, pos=pos)
 
-    def parse_term__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term]]:
+    def parse_term__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term] | None:
         result: fltk_cst.Term = fltk_cst.Term(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_term__alt0__item0(pos=pos):
             pos = item0.pos
@@ -594,12 +558,10 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_term__alt0__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier]]:
+    def parse_term__alt0__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier] | None:
         return self.apply__parse_identifier(pos=pos)
 
-    def parse_term__alt1(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term]]:
+    def parse_term__alt1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term] | None:
         result: fltk_cst.Term = fltk_cst.Term(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_term__alt1__item0(pos=pos):
             pos = item0.pos
@@ -609,12 +571,10 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_term__alt1__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Literal]]:
+    def parse_term__alt1__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Literal] | None:
         return self.apply__parse_literal(pos=pos)
 
-    def parse_term__alt2(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term]]:
+    def parse_term__alt2(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term] | None:
         result: fltk_cst.Term = fltk_cst.Term(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_term__alt2__item0(pos=pos):
             pos = item0.pos
@@ -634,20 +594,18 @@ class Parser:
 
     def parse_term__alt2__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="/")
 
-    def parse_term__alt2__item1(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.RawString]]:
+    def parse_term__alt2__item1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.RawString] | None:
         return self.apply__parse_raw_string(pos=pos)
 
     def parse_term__alt2__item2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="/")
 
-    def parse_term__alt3(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term]]:
+    def parse_term__alt3(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Term] | None:
         result: fltk_cst.Term = fltk_cst.Term(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_term__alt3__item0(pos=pos):
             pos = item0.pos
@@ -671,22 +629,18 @@ class Parser:
 
     def parse_term__alt3__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="(")
 
-    def parse_term__alt3__item1(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives]]:
+    def parse_term__alt3__item1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Alternatives] | None:
         return self.apply__parse_alternatives(pos=pos)
 
     def parse_term__alt3__item2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal=")")
 
-    def parse_disposition(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition]]:
+    def parse_disposition(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition] | None:
         if alt0 := self.parse_disposition__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_disposition__alt1(pos=pos):
@@ -695,16 +649,12 @@ class Parser:
             return alt2
         return None
 
-    def apply__parse_disposition(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition]]:
+    def apply__parse_disposition(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition] | None:
         return self.packrat.apply(
             rule_callable=self.parse_disposition, rule_id=6, rule_cache=self._cache__parse_disposition, pos=pos
         )
 
-    def parse_disposition__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition]]:
+    def parse_disposition__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition] | None:
         result: fltk_cst.Disposition = fltk_cst.Disposition(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_disposition__alt0__item0(pos=pos):
             pos = item0.pos
@@ -716,12 +666,10 @@ class Parser:
 
     def parse_disposition__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="%")
 
-    def parse_disposition__alt1(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition]]:
+    def parse_disposition__alt1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition] | None:
         result: fltk_cst.Disposition = fltk_cst.Disposition(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_disposition__alt1__item0(pos=pos):
             pos = item0.pos
@@ -733,12 +681,10 @@ class Parser:
 
     def parse_disposition__alt1__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="$")
 
-    def parse_disposition__alt2(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition]]:
+    def parse_disposition__alt2(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Disposition] | None:
         result: fltk_cst.Disposition = fltk_cst.Disposition(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_disposition__alt2__item0(pos=pos):
             pos = item0.pos
@@ -750,10 +696,10 @@ class Parser:
 
     def parse_disposition__alt2__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="!")
 
-    def parse_quantifier(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier]]:
+    def parse_quantifier(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier] | None:
         if alt0 := self.parse_quantifier__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_quantifier__alt1(pos=pos):
@@ -762,16 +708,12 @@ class Parser:
             return alt2
         return None
 
-    def apply__parse_quantifier(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier]]:
+    def apply__parse_quantifier(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier] | None:
         return self.packrat.apply(
             rule_callable=self.parse_quantifier, rule_id=7, rule_cache=self._cache__parse_quantifier, pos=pos
         )
 
-    def parse_quantifier__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier]]:
+    def parse_quantifier__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier] | None:
         result: fltk_cst.Quantifier = fltk_cst.Quantifier(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_quantifier__alt0__item0(pos=pos):
             pos = item0.pos
@@ -783,12 +725,10 @@ class Parser:
 
     def parse_quantifier__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="?")
 
-    def parse_quantifier__alt1(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier]]:
+    def parse_quantifier__alt1(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier] | None:
         result: fltk_cst.Quantifier = fltk_cst.Quantifier(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_quantifier__alt1__item0(pos=pos):
             pos = item0.pos
@@ -800,12 +740,10 @@ class Parser:
 
     def parse_quantifier__alt1__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="+")
 
-    def parse_quantifier__alt2(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier]]:
+    def parse_quantifier__alt2(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Quantifier] | None:
         result: fltk_cst.Quantifier = fltk_cst.Quantifier(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_quantifier__alt2__item0(pos=pos):
             pos = item0.pos
@@ -817,24 +755,20 @@ class Parser:
 
     def parse_quantifier__alt2__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="*")
 
-    def parse_identifier(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier]]:
+    def parse_identifier(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier] | None:
         if alt0 := self.parse_identifier__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_identifier(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier]]:
+    def apply__parse_identifier(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier] | None:
         return self.packrat.apply(
             rule_callable=self.parse_identifier, rule_id=8, rule_cache=self._cache__parse_identifier, pos=pos
         )
 
-    def parse_identifier__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier]]:
+    def parse_identifier__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Identifier] | None:
         result: fltk_cst.Identifier = fltk_cst.Identifier(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_identifier__alt0__item0(pos=pos):
             pos = item0.pos
@@ -846,24 +780,20 @@ class Parser:
 
     def parse_identifier__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_regex(pos=pos, regex="[_a-z][_a-z0-9]*")
 
-    def parse_raw_string(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.RawString]]:
+    def parse_raw_string(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.RawString] | None:
         if alt0 := self.parse_raw_string__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_raw_string(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.RawString]]:
+    def apply__parse_raw_string(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.RawString] | None:
         return self.packrat.apply(
             rule_callable=self.parse_raw_string, rule_id=9, rule_cache=self._cache__parse_raw_string, pos=pos
         )
 
-    def parse_raw_string__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.RawString]]:
+    def parse_raw_string__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.RawString] | None:
         result: fltk_cst.RawString = fltk_cst.RawString(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_raw_string__alt0__item0(pos=pos):
             pos = item0.pos
@@ -875,22 +805,20 @@ class Parser:
 
     def parse_raw_string__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_regex(pos=pos, regex="([^\\/\\n\\\\]|\\\\.)+")
 
-    def parse_literal(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Literal]]:
+    def parse_literal(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Literal] | None:
         if alt0 := self.parse_literal__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_literal(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Literal]]:
+    def apply__parse_literal(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Literal] | None:
         return self.packrat.apply(
             rule_callable=self.parse_literal, rule_id=10, rule_cache=self._cache__parse_literal, pos=pos
         )
 
-    def parse_literal__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Literal]]:
+    def parse_literal__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Literal] | None:
         result: fltk_cst.Literal = fltk_cst.Literal(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_literal__alt0__item0(pos=pos):
             pos = item0.pos
@@ -902,20 +830,20 @@ class Parser:
 
     def parse_literal__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
-        return self.consume_regex(pos=pos, regex='("([^"\\n\\\\]|\\\\.)+"|\'([^\'\\n\\\\]|\\\\.)+\')')
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
+        return self.consume_regex(pos=pos, regex="(\"([^\"\\n\\\\]|\\\\.)+\"|'([^'\\n\\\\]|\\\\.)+')")
 
-    def parse_trivia(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia]]:
+    def parse_trivia(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia] | None:
         if alt0 := self.parse_trivia__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_trivia(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia]]:
+    def apply__parse_trivia(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia] | None:
         return self.packrat.apply(
             rule_callable=self.parse_trivia, rule_id=11, rule_cache=self._cache__parse_trivia, pos=pos
         )
 
-    def parse_trivia__alt0(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia]]:
+    def parse_trivia__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia] | None:
         result: fltk_cst.Trivia = fltk_cst.Trivia(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_trivia__alt0__item0(pos=pos):
             pos = item0.pos
@@ -927,7 +855,7 @@ class Parser:
 
     def parse_trivia__alt0__item0__alts(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia] | None:
         if alt0 := self.parse_trivia__alt0__item0__alts__alt0(pos=pos):
             return alt0
         if alt1 := self.parse_trivia__alt0__item0__alts__alt1(pos=pos):
@@ -938,7 +866,7 @@ class Parser:
 
     def parse_trivia__alt0__item0__alts__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia] | None:
         result: fltk_cst.Trivia = fltk_cst.Trivia(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_trivia__alt0__item0__alts__alt0__item0(pos=pos):
             pos = item0.pos
@@ -950,12 +878,12 @@ class Parser:
 
     def parse_trivia__alt0__item0__alts__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Whitespace]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Whitespace] | None:
         return self.apply__parse_whitespace(pos=pos)
 
     def parse_trivia__alt0__item0__alts__alt1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia] | None:
         result: fltk_cst.Trivia = fltk_cst.Trivia(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_trivia__alt0__item0__alts__alt1__item0(pos=pos):
             pos = item0.pos
@@ -967,12 +895,12 @@ class Parser:
 
     def parse_trivia__alt0__item0__alts__alt1__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.LineComment]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.LineComment] | None:
         return self.apply__parse_line_comment(pos=pos)
 
     def parse_trivia__alt0__item0__alts__alt2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia] | None:
         result: fltk_cst.Trivia = fltk_cst.Trivia(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_trivia__alt0__item0__alts__alt2__item0(pos=pos):
             pos = item0.pos
@@ -984,12 +912,10 @@ class Parser:
 
     def parse_trivia__alt0__item0__alts__alt2__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.BlockComment]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.BlockComment] | None:
         return self.apply__parse_block_comment(pos=pos)
 
-    def parse_trivia__alt0__item0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia]]:
+    def parse_trivia__alt0__item0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Trivia] | None:
         result: fltk_cst.Trivia = fltk_cst.Trivia(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         while one_result := self.parse_trivia__alt0__item0__alts(pos=pos):
             pos = one_result.pos
@@ -999,21 +925,17 @@ class Parser:
         result.span = fltk.fegen.pyrt.terminalsrc.Span(start=result.span.start, end=pos)
         return fltk.fegen.pyrt.memo.ApplyResult(pos=pos, result=result)
 
-    def parse_whitespace(self, pos: int) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Whitespace]]:
+    def parse_whitespace(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Whitespace] | None:
         if alt0 := self.parse_whitespace__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_whitespace(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Whitespace]]:
+    def apply__parse_whitespace(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Whitespace] | None:
         return self.packrat.apply(
             rule_callable=self.parse_whitespace, rule_id=12, rule_cache=self._cache__parse_whitespace, pos=pos
         )
 
-    def parse_whitespace__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Whitespace]]:
+    def parse_whitespace__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.Whitespace] | None:
         result: fltk_cst.Whitespace = fltk_cst.Whitespace(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_whitespace__alt0__item0(pos=pos):
             pos = item0.pos
@@ -1025,26 +947,20 @@ class Parser:
 
     def parse_whitespace__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_regex(pos=pos, regex="\\s+")
 
-    def parse_line_comment(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.LineComment]]:
+    def parse_line_comment(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.LineComment] | None:
         if alt0 := self.parse_line_comment__alt0(pos=pos):
             return alt0
         return None
 
-    def apply__parse_line_comment(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.LineComment]]:
+    def apply__parse_line_comment(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.LineComment] | None:
         return self.packrat.apply(
             rule_callable=self.parse_line_comment, rule_id=13, rule_cache=self._cache__parse_line_comment, pos=pos
         )
 
-    def parse_line_comment__alt0(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.LineComment]]:
+    def parse_line_comment__alt0(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.LineComment] | None:
         result: fltk_cst.LineComment = fltk_cst.LineComment(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_line_comment__alt0__item0(pos=pos):
             pos = item0.pos
@@ -1066,36 +982,34 @@ class Parser:
 
     def parse_line_comment__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="//")
 
     def parse_line_comment__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_regex(pos=pos, regex="[^\\n]*")
 
     def parse_line_comment__alt0__item2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="\n")
 
-    def parse_block_comment(
-        self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.BlockComment]]:
+    def parse_block_comment(self, pos: int) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.BlockComment] | None:
         if alt0 := self.parse_block_comment__alt0(pos=pos):
             return alt0
         return None
 
     def apply__parse_block_comment(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.BlockComment]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.BlockComment] | None:
         return self.packrat.apply(
             rule_callable=self.parse_block_comment, rule_id=14, rule_cache=self._cache__parse_block_comment, pos=pos
         )
 
     def parse_block_comment__alt0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.BlockComment]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk_cst.BlockComment] | None:
         result: fltk_cst.BlockComment = fltk_cst.BlockComment(span=fltk.fegen.pyrt.terminalsrc.Span(start=pos, end=-1))
         if item0 := self.parse_block_comment__alt0__item0(pos=pos):
             pos = item0.pos
@@ -1121,15 +1035,15 @@ class Parser:
 
     def parse_block_comment__alt0__item0(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="/*")
 
     def parse_block_comment__alt0__item1(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_regex(pos=pos, regex="[^*]*(?:\\*(?!\\/)[^*]*)*")
 
     def parse_block_comment__alt0__item2(
         self, pos: int
-    ) -> typing.Optional[fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span]]:
+    ) -> fltk.fegen.pyrt.memo.ApplyResult[int, fltk.fegen.pyrt.terminalsrc.Span] | None:
         return self.consume_literal(pos=pos, literal="*/")
