@@ -29,7 +29,7 @@ class Cst2Gsm:
     def visit_items(self, items: cst.Items) -> gsm.Items:
         gsm_items = []
         sep_after = []
-        for (item_label, item), (sep_label, _) in zip(items.children[::2], items.children[1::2]):
+        for (item_label, item), (sep_label, _) in zip(items.children[::2], items.children[1::2], strict=False):
             assert item_label == cst.Items.Label.ITEM and isinstance(item, cst.Item)  # noqa: S101
             gsm_items.append(self.visit_item(item))
             if sep_label == cst.Items.Label.WS_REQUIRED:
